refactor: made AddCardModal as a general container and fixed the condition for mounting cards
